Update: Modified test code due to refactoring.
