+++ src/main/kotlin/com/musinsa/musinsacodingassignment/product/entity/ProductEntity.kt
@@ -1,7 +1,7 @@
 package com.musinsa.musinsacodingassignment.product.entity
 
 import com.musinsa.musinsacodingassignment.brand.entity.BrandEntity
-import com.musinsa.musinsacodingassignment.brand.entity.CategoryEntity
+import com.musinsa.musinsacodingassignment.category.entity.CategoryEntity
 import com.musinsa.musinsacodingassignment.common.domain.BaseEntity
 import jakarta.persistence.*
 
@@ -20,7 +20,7 @@ class ProductEntity(
     @Column(nullable = false)
     val price: Int
 ) : BaseEntity() {
-    constructor(id: Long,  brand: BrandEntity, category: CategoryEntity, price: Int) : this(
+    constructor(id: Long, brand: BrandEntity, category: CategoryEntity, price: Int) : this(
         brand,
         category,
         price
