+++ src/main/kotlin/com/musinsa/musinsacodingassignment/product/repository/ProductRepository.kt
@@ -1,13 +1,11 @@
 package com.musinsa.musinsacodingassignment.product.repository
 
 import com.musinsa.musinsacodingassignment.brand.entity.BrandEntity
-import com.musinsa.musinsacodingassignment.brand.entity.CategoryEntity
-import com.musinsa.musinsacodingassignment.category.domain.Category
+import com.musinsa.musinsacodingassignment.category.entity.CategoryEntity
 import com.musinsa.musinsacodingassignment.product.entity.ProductEntity
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Modifying
 import org.springframework.data.jpa.repository.Query
-import org.springframework.transaction.annotation.Transactional
 import java.time.LocalDateTime
 
 interface ProductRepository : JpaRepository<ProductEntity, Long> {
